+++ packages/babel-helper-transform-fixture-test-runner/src/index.js
@@ -39,6 +39,7 @@ vm.runInContext(
 // Initialize the test context with the polyfill, and then freeze the global to prevent implicit
 // global creation in tests, which could cause things to bleed between tests.
 runModuleInTestContext("babel-polyfill", __filename);
+runModuleInTestContext("babel-polyfill-bigint", __filename);
 
 // Populate the "babelHelpers" global with Babel's helper utilities.
 runCodeInTestContext(buildExternalHelpers());
