+++ packages/babel-polyfill-bigint/.npmignore
@@ -1,0 +1,3 @@
+src
+test
+*.log
